--- shucks-app/src/App/App.css
@@ -1,3 +1,0 @@
-.App {
-  margin: 0.5em;
-}
