--- shucks-app/src/Message/Message.css
@@ -1,23 +1,0 @@
-.Message * {
-  margin: 0.5em;
-  padding: 1em;
-  max-width: 60%;
-}
-
-.recieved {
-  float: left;
-  color: #000000;
-  background-color: #eeeeee;
-  border-radius: 1.5em 1.5em 1.5em 0.2em;
-}
-
-.sent {
-  margin: 0.5em;
-  padding: 1em;
-  max-width: 60%;
-
-  float: right;
-  color: #ffffff;
-  background-color: #8844dd;
-  border-radius: 1.5em 1.5em 0.2em 1.5em;
-}
