+++ shucks-app/src/components/App/App.test.js
