+++ shucks-app/src/components/Message/Message.css
@@ -1,0 +1,37 @@
+.messageBubble {
+  margin: 1em 1em 0.1em 1em;
+  padding: 1em;
+  max-width: 60%;
+  position:relative;
+}
+
+.recieved .messageBubble {
+  float:left;
+  color: #000000;
+  background-color: #eeeeee;
+  border-radius: 1.5em 1.5em 1.5em 0.2em;
+}
+
+.sent .messageBubble {
+  float:right;
+  color: #ffffff;
+  background-color: #8844dd;
+  border-radius: 1.5em 1.5em 0.2em 1.5em;
+}
+
+.userInfo {
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+  font-size: 9pt;
+  margin: 0.1em 1em 1em 1em;
+  color: #555555;
+}
+
+.recieved .userInfo {
+  margin-left: 1.5em;
+  float:left;
+}
+
+.sent .userInfo {
+  margin-right: 1.5em;
+  float:right;
+}
