+++ shucks-app/src/components/Message/Message.js
@@ -1,0 +1,27 @@
+import React, { Component } from 'react';
+import './Message.css';
+
+class Message extends Component {
+  constructor(props) {
+    super(props);
+    this.username = props.username
+  }
+  render() {
+    return (
+      <div className="Message">
+        <div className={this.props.username == "orwinmc" ? "sent" : "recieved"}>
+          <div className="messageBubble">
+            {this.props.msg}
+          </div>
+          <br clear="all" />
+          <div className="userInfo">
+            {this.props.username}
+          </div>
+          <br clear="all" />
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Message;
