+++ shucks-app/src/components/MessageBoard/MessageBoard.css
@@ -1,0 +1,2 @@
+.MessageBoard {
+}
